+++ tests/testthat/test-learner_rfsrc_s4.R
@@ -6,7 +6,7 @@ testthat::skip_on_cran()
 
 familiar:::test_hyperparameter_optimisation(learners=familiar:::.get_available_rfsrc_learners(show_general=TRUE),
                                             debug=FALSE,
-                                            parallel=FALSE)
+                                            parallel=TRUE)
 
 familiar:::test_all_learners_train_predict_vimp(learners=familiar:::.get_available_rfsrc_learners(show_general=FALSE),
                                                 hyperparameter_list=list("count"=list("n_tree"=4,
@@ -75,7 +75,7 @@ testthat::test_that("Random forest SRC model trained correctly", {
 testthat::test_that("Random forest SRC model has variable importance", {
   
   # Extract the variable importance table.
-  vimp_table <- familiar:::..vimp(good_model)
+  vimp_table <- familiar:::..vimp(good_model, data=good_data)
   
   # Expect that the vimp table has two rows.
   testthat::expect_equal(nrow(vimp_table), 13)
@@ -96,7 +96,7 @@ testthat::test_that("Random forest SRC model can train on wide data", {
   testthat::expect_equal(familiar:::model_is_trained(wide_model), TRUE)
   
   # Variable importance table is present.
-  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model)), FALSE)
+  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model, data=wide_data)), FALSE)
   
   # Valid predictions.
   testthat::expect_equal(familiar:::any_predictions_valid(familiar:::.predict(wide_model, wide_data), outcome_type=wide_data@outcome_type), TRUE)
@@ -142,7 +142,7 @@ testthat::test_that("Random forest SRC model trained correctly", {
 testthat::test_that("Random forest SRC model has variable importance", {
   
   # Extract the variable importance table.
-  vimp_table <- familiar:::..vimp(good_model)
+  vimp_table <- familiar:::..vimp(good_model, data=good_data)
   
   # Expect that the vimp table has two rows.
   testthat::expect_equal(nrow(vimp_table), 10)
@@ -161,7 +161,7 @@ testthat::test_that("Random forest SRC model can train on wide data", {
   testthat::expect_equal(familiar:::model_is_trained(wide_model), TRUE)
   
   # Variable importance table is present.
-  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model)), FALSE)
+  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model, data=wide_data)), FALSE)
   
   # Valid predictions.
   testthat::expect_equal(familiar:::any_predictions_valid(familiar:::.predict(wide_model, wide_data), outcome_type=wide_data@outcome_type), TRUE)
@@ -207,7 +207,7 @@ testthat::test_that("Random forest SRC model trained correctly", {
 testthat::test_that("Random forest SRC model has variable importance", {
   
   # Extract the variable importance table.
-  vimp_table <- familiar:::..vimp(good_model)
+  vimp_table <- familiar:::..vimp(good_model, data=good_data)
   
   # Expect that the vimp table has two rows.
   testthat::expect_equal(nrow(vimp_table), 8)
@@ -227,7 +227,7 @@ testthat::test_that("Random forest SRC model can train on wide data", {
   testthat::expect_equal(familiar:::model_is_trained(wide_model), TRUE)
   
   # Variable importance table is present.
-  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model)), FALSE)
+  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model, data=wide_data)), FALSE)
   
   # Valid predictions.
   testthat::expect_equal(familiar:::any_predictions_valid(familiar:::.predict(wide_model, wide_data), outcome_type=wide_data@outcome_type), TRUE)
@@ -274,7 +274,7 @@ testthat::test_that("Random forest SRC model trained correctly", {
 testthat::test_that("Random forest SRC model has variable importance", {
   
   # Extract the variable importance table.
-  vimp_table <- familiar:::..vimp(good_model)
+  vimp_table <- familiar:::..vimp(good_model, data=good_data)
   
   # Expect that the vimp table has two rows.
   testthat::expect_equal(nrow(vimp_table), 4)
@@ -295,7 +295,7 @@ testthat::test_that("Random forest SRC model can train on wide data", {
   testthat::expect_equal(familiar:::model_is_trained(wide_model), TRUE)
   
   # Variable importance table is present.
-  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model)), FALSE)
+  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model, data=wide_data)), FALSE)
   
   # Valid predictions.
   testthat::expect_equal(familiar:::any_predictions_valid(familiar:::.predict(wide_model, wide_data), outcome_type=wide_data@outcome_type), TRUE)
@@ -351,7 +351,7 @@ testthat::test_that("Random forest SRC model trained correctly", {
 testthat::test_that("Random forest SRC model has variable importance", {
   
   # Extract the variable importance table.
-  vimp_table <- familiar:::..vimp(good_model)
+  vimp_table <- familiar:::..vimp(good_model, data=good_data)
   
   # Expect that the vimp table has three rows.
   testthat::expect_equal(nrow(vimp_table), 3)
@@ -371,7 +371,7 @@ testthat::test_that("Random forest SRC model can train on wide data", {
   testthat::expect_equal(familiar:::model_is_trained(wide_model), TRUE)
   
   # Variable importance table is present
-  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model)), FALSE)
+  testthat::expect_equal(is_empty(familiar:::..vimp(wide_model, data=wide_data)), FALSE)
   
   # Valid predictions are possible.
   testthat::expect_equal(familiar:::any_predictions_valid(familiar:::.predict(wide_model, wide_data), outcome_type=wide_data@outcome_type), TRUE)