+++ tests/testthat/test-vimp_rfsrc_S4.R
@@ -4,11 +4,11 @@ familiar:::test_all_vimp_methods_available(familiar:::.get_available_rfsrc_vimp_
 testthat::skip_on_cran()
 
 familiar:::test_hyperparameter_optimisation(vimp_methods=familiar:::.get_available_rfsrc_vimp_methods(show_general=TRUE),
-                                            debug=TRUE,
+                                            debug=FALSE,
                                             parallel=FALSE)
 
 familiar:::test_all_vimp_methods(familiar:::.get_available_rfsrc_vimp_methods(show_general=FALSE),
-                                 debug=TRUE,
+                                 debug=FALSE,
                                  hyperparameter_list=list("count"=list("n_tree"=4,
                                                                        "sample_size"=0.50,
                                                                        "m_try"=0.3,